# Improve `vec_equal_na()` performance with data frames