--- a/src/equal.c
+++ b/src/equal.c
@@ -527,46 +527,59 @@ int equal_na(SEXP x, R_len_t i) {
 
 #define EQUAL_NA(CTYPE, CONST_DEREF, SCALAR_EQUAL_NA)     \
   do {                                                    \
-    const CTYPE* xp = CONST_DEREF(x);                     \
+    SEXP out = PROTECT(Rf_allocVector(LGLSXP, size));     \
+    int* p_out = LOGICAL(out);                            \
                                                           \
-    for (R_len_t i = 0; i < n; ++i, ++xp) {               \
-      p[i] = SCALAR_EQUAL_NA(xp);                         \
+    const CTYPE* p_x = CONST_DEREF(x);                    \
+                                                          \
+    for (R_len_t i = 0; i < size; ++i, ++p_x) {           \
+      p_out[i] = SCALAR_EQUAL_NA(p_x);                    \
     }                                                     \
+                                                          \
+    UNPROTECT(2);                                         \
+    return out;                                           \
   }                                                       \
   while (0)
 
 #define EQUAL_NA_BARRIER(SCALAR_EQUAL_NA)                 \
   do {                                                    \
-    for (R_len_t i = 0; i < n; ++i) {                     \
-      p[i] = SCALAR_EQUAL_NA(x, i);                       \
+    SEXP out = PROTECT(Rf_allocVector(LGLSXP, size));     \
+    int* p_out = LOGICAL(out);                            \
+                                                          \
+    for (R_len_t i = 0; i < size; ++i) {                  \
+      p_out[i] = SCALAR_EQUAL_NA(x, i);                   \
     }                                                     \
+                                                          \
+    UNPROTECT(2);                                         \
+    return out;                                           \
   }                                                       \
   while (0)
 
+static SEXP df_equal_na(SEXP x, R_len_t n_row);
+
 // [[ register() ]]
 SEXP vctrs_equal_na(SEXP x) {
-  R_len_t n = vec_size(x);
-  SEXP out = PROTECT(Rf_allocVector(LGLSXP, n));
-  int32_t* p = LOGICAL(out);
+  R_len_t size = vec_size(x);
 
   x = PROTECT(vec_proxy_equal(x));
 
   enum vctrs_type type = vec_proxy_typeof(x);
 
   switch (type) {
-  case vctrs_type_logical:   EQUAL_NA(int, LOGICAL_RO, lgl_equal_na_scalar); break;
-  case vctrs_type_integer:   EQUAL_NA(int, INTEGER_RO, int_equal_na_scalar); break;
-  case vctrs_type_double:    EQUAL_NA(double, REAL_RO, dbl_equal_na_scalar); break;
-  case vctrs_type_complex:   EQUAL_NA(Rcomplex, COMPLEX_RO, cpl_equal_na_scalar); break;
-  case vctrs_type_character: EQUAL_NA(SEXP, STRING_PTR_RO, chr_equal_na_scalar); break;
-  case vctrs_type_list:      EQUAL_NA_BARRIER(list_equal_na_scalar); break;
-  case vctrs_type_dataframe: EQUAL_NA_BARRIER(df_equal_na_scalar); break;
+  case vctrs_type_logical:   EQUAL_NA(int, LOGICAL_RO, lgl_equal_na_scalar);
+  case vctrs_type_integer:   EQUAL_NA(int, INTEGER_RO, int_equal_na_scalar);
+  case vctrs_type_double:    EQUAL_NA(double, REAL_RO, dbl_equal_na_scalar);
+  case vctrs_type_complex:   EQUAL_NA(Rcomplex, COMPLEX_RO, cpl_equal_na_scalar);
+  case vctrs_type_character: EQUAL_NA(SEXP, STRING_PTR_RO, chr_equal_na_scalar);
+  case vctrs_type_list:      EQUAL_NA_BARRIER(list_equal_na_scalar);
+  case vctrs_type_dataframe: {
+    SEXP out = df_equal_na(x, size);
+    UNPROTECT(1);
+    return out;
+  }
   case vctrs_type_scalar:    Rf_errorcall(R_NilValue, "Can't detect `NA` values in scalars with `vctrs_equal_na()`.");
   default:                   Rf_error("Unimplemented type in `vctrs_equal_na()`.");
   }
-
-  UNPROTECT(2);
-  return out;
 }
 
 #undef EQUAL_NA
@@ -610,3 +623,112 @@ static int df_equal_na_scalar(SEXP x, R_len_t i) {
 
   return true;
 }
+
+// -----------------------------------------------------------------------------
+
+static struct vctrs_df_rowwise_info vec_equal_na_col(SEXP x,
+                                                     struct vctrs_df_rowwise_info info,
+                                                     R_len_t n_row);
+
+static struct vctrs_df_rowwise_info df_equal_na_impl(SEXP x,
+                                                     struct vctrs_df_rowwise_info info,
+                                                     R_len_t n_row) {
+  int n_col = Rf_length(x);
+
+  for (R_len_t i = 0; i < n_col; ++i) {
+    SEXP col = VECTOR_ELT(x, i);
+
+    info = vec_equal_na_col(col, info, n_row);
+
+    // If we know all comparison values, break
+    if (info.remaining == 0) {
+      break;
+    }
+  }
+
+  return info;
+}
+
+static SEXP df_equal_na(SEXP x, R_len_t n_row) {
+  int nprot = 0;
+
+  // Same rowwise info as `df_equal()`
+  struct vctrs_df_rowwise_info info = init_rowwise_equal_info(n_row);
+  PROTECT_DF_ROWWISE_INFO(&info, &nprot);
+
+  info = df_equal_na_impl(x, info, n_row);
+
+  UNPROTECT(nprot);
+  return info.out;
+}
+
+// -----------------------------------------------------------------------------
+
+#define EQUAL_NA_COL(CTYPE, CONST_DEREF, SCALAR_EQUAL_NA) \
+do {                                                      \
+  int* p_out = LOGICAL(info.out);                         \
+                                                          \
+  const CTYPE* p_x = CONST_DEREF(x);                      \
+                                                          \
+  for (R_len_t i = 0; i < n_row; ++i, ++p_x) {            \
+    if (info.p_row_known[i]) {                            \
+      continue;                                           \
+    }                                                     \
+                                                          \
+    if (!SCALAR_EQUAL_NA(p_x)) {                          \
+      p_out[i] = 0;                                       \
+      info.p_row_known[i] = true;                         \
+      --info.remaining;                                   \
+                                                          \
+      if (info.remaining == 0) {                          \
+        break;                                            \
+      }                                                   \
+    }                                                     \
+  }                                                       \
+                                                          \
+  return info;                                            \
+}                                                         \
+while (0)
+
+#define EQUAL_NA_COL_BARRIER(SCALAR_EQUAL_NA) \
+do {                                          \
+  int* p_out = LOGICAL(info.out);             \
+                                              \
+  for (R_len_t i = 0; i < n_row; ++i) {       \
+    if (info.p_row_known[i]) {                \
+      continue;                               \
+    }                                         \
+                                              \
+    if (!SCALAR_EQUAL_NA(x, i)) {             \
+      p_out[i] = 0;                           \
+      info.p_row_known[i] = true;             \
+      --info.remaining;                       \
+                                              \
+      if (info.remaining == 0) {              \
+        break;                                \
+      }                                       \
+    }                                         \
+  }                                           \
+                                              \
+  return info;                                \
+}                                             \
+while (0)
+
+static struct vctrs_df_rowwise_info vec_equal_na_col(SEXP x,
+                                                     struct vctrs_df_rowwise_info info,
+                                                     R_len_t n_row) {
+  switch (vec_proxy_typeof(x)) {
+  case vctrs_type_logical:   EQUAL_NA_COL(int, LOGICAL_RO, lgl_equal_na_scalar);
+  case vctrs_type_integer:   EQUAL_NA_COL(int, INTEGER_RO, int_equal_na_scalar);
+  case vctrs_type_double:    EQUAL_NA_COL(double, REAL_RO, dbl_equal_na_scalar);
+  case vctrs_type_complex:   EQUAL_NA_COL(Rcomplex, COMPLEX_RO, cpl_equal_na_scalar);
+  case vctrs_type_character: EQUAL_NA_COL(SEXP, STRING_PTR_RO, chr_equal_na_scalar);
+  case vctrs_type_list:      EQUAL_NA_COL_BARRIER(list_equal_na_scalar);
+  case vctrs_type_dataframe: return df_equal_na_impl(x, info, n_row);
+  case vctrs_type_scalar:    Rf_errorcall(R_NilValue, "Can't compare scalars with `vec_equal_na()`");
+  default:                   Rf_error("Unimplemented type in `vec_equal_na()`");
+  }
+}
+
+#undef EQUAL_NA_COL
+#undef EQUAL_NA_COL_BARRIER