This PR updates `vec_equal_na()` to use the faster traversal method with data frames (looping over the columns one at a time, and early exiting if we know all the results).

It also adds support for `vec_equal_na()` with raw vectors. This always returns `FALSE`, but it doesn't error now.

This worst case scenario below only happens when all elements are `NA`. Even that is much faster than the current master implementation.

```r
library(vctrs)

# worst case
x <- data.frame(x = rep(NA, 1e5), y = rep(NA, 1e5), z = rep(NA, 1e5))

bench::mark(vec_equal_na(x))

# master
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal_na(x)   2.57ms   2.75ms      349.     395KB     2.14

# this PR
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal_na(x)    300µs    358µs     2596.     492KB     22.7
```

```r
# short circuit after the first column
y <- data.frame(x = rep(1, 1e5), y = rep(2, 1e5), z = rep(3, 1e5))

bench::mark(vec_equal_na(y))

# master
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal_na(y)    894µs    953µs     1017.     391KB     6.24

# this PR
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal_na(y)    117µs    150µs     6091.     488KB     54.7
```