@DavisVaughan 2020-03-10
@lionel- do you want more information than this?

I sort of feel like this should actually be called `vctrs_df_colwise_info` now, since we iterate over columns, but I'm not sure

Should this move to `colwise.h`?

https://github.com/r-lib/vctrs/blob/97c6fb4cd7ef0828ac299328d48c9515462140e6/src/vctrs.h#L473-L505