@DavisVaughan 2020-03-10
> I imagine vec-complete-cases will work similarly but short-circuit as soon as it finds a missing value?

Yes, the code will be heavily duplicated, but I don't mind it that much