@lionel- 2020-03-10
> do you want more information than this?

Ah no, that's good.

> I sort of feel like this should actually be called vctrs_colwise_info now, since we iterate over columns, but I'm not sure

Maybe something along the lines of `df_short_circuit_info`?