@DavisVaughan 2020-03-10
I'm not sure `df_short_circuit_info` conveys exactly the right meaning because it carries the result object, `out`, as well, so its not just about short circuiting with `row_known` and `remaining`. I went with `df_col_traversal_info`.