@DavisVaughan 2020-03-10
I have an idea to slightly simplify the info object, but I'll leave it and the renaming for a separate PR